Formatted files and rename components/OnboardingApplication to components/employee-info-forms
